requirements.txt
flask==2.3.3
flask-login==0.6.2
werkzeug==2.3.7
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pillow==10.0.0 